backend/pyproject.toml
[project]
name = "dejaship"
version = "0.1.0"
description = "The Global Intent Ledger for AI Agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "fastembed>=0.4.0",
    "mcp>=1.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "anyio>=4.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "instructor>=1.0.0",
    "openai>=1.0.0",
    "testcontainers[postgres]>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dejaship"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "agent_sim: deterministic agent simulation and fixture replay tests",
    "live_llm: tests that call a real external LLM provider",
    "slow: intentionally slower end-to-end simulations",
]
